Move helper function from case_description module to case_description_attributes
