+++ tests/alfacase/test_case_description.py
@@ -11,15 +11,15 @@ from ..common_testing.alfasim_sdk_common_testing.case_builders import (
 )
 from _alfasim_sdk import constants
 from _alfasim_sdk.alfacase import case_description
-from _alfasim_sdk.alfacase.case_description import attrib_enum
-from _alfasim_sdk.alfacase.case_description import attrib_instance
-from _alfasim_sdk.alfacase.case_description import attrib_instance_list
-from _alfasim_sdk.alfacase.case_description import attrib_scalar
-from _alfasim_sdk.alfacase.case_description import collapse_array_repr
 from _alfasim_sdk.alfacase.case_description import InvalidReferenceError
 from _alfasim_sdk.alfacase.case_description import MaterialDescription
-from _alfasim_sdk.alfacase.case_description import numpy_array_validator
 from _alfasim_sdk.alfacase.case_description import PvtModelTableParametersDescription
+from _alfasim_sdk.alfacase.case_description_attributes import attrib_enum
+from _alfasim_sdk.alfacase.case_description_attributes import attrib_instance
+from _alfasim_sdk.alfacase.case_description_attributes import attrib_instance_list
+from _alfasim_sdk.alfacase.case_description_attributes import attrib_scalar
+from _alfasim_sdk.alfacase.case_description_attributes import collapse_array_repr
+from _alfasim_sdk.alfacase.case_description_attributes import numpy_array_validator
 from _alfasim_sdk.constants import NodeCellType
